backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.13
python-dotenv==1.0.1
PyJWT==2.10.1
pytest==8.4.2
werkzeug==3.1.3
